Propose initial refactor
